+++ cmd/tx-decoder/app.go
@@ -78,7 +78,7 @@ func initConsumerGroup(ctx context.Context) {
 
 	// Generic handlers on every worker
 	engine.Register(opentracing.GlobalHandler())
-	engine.Register(logger.Logger)
+	engine.Register(logger.Logger("debug"))
 	engine.Register(sarama.Loader)
 	engine.Register(offset.Marker)
 	engine.Register(opentracing.GlobalHandler())