+++ cmd/tx-sender/app.go
@@ -72,7 +72,7 @@ func initComponents(ctx context.Context) {
 func registerHandlers() {
 	// Generic handlers on every worker
 	engine.Register(opentracing.GlobalHandler())
-	engine.Register(logger.Logger)
+	engine.Register(logger.Logger("info"))
 	engine.Register(sarama.Loader)
 	engine.Register(offset.Marker)
 	engine.Register(opentracing.GlobalHandler())