+++ pkg/server/grpc/server_test.go
@@ -2,6 +2,6 @@ package grpcserver
 
 import "testing"
 
-func TestNewServerWithDefaultOptions(t *testing.T) {
-	NewServerWithDefaultOptions()
+func TestNewServer(t *testing.T) {
+	NewServer()
 }